moving etasu endpoints into their own middleware component
